+++ Source/GenerateTypeScript/Program.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using System.Reflection;
 using System.Text;
 using System.Text.RegularExpressions;
-using Microsoft.AspNetCore.Mvc;
 using MoreLinq;
 
 namespace GenerateTypeScript